fix - all worked

+++ feature-main-screen/src/main/java/com/example/feature_main_screen/domain/models/StageInfo.kt
@@ -3,5 +3,5 @@ package com.example.feature_main_screen.domain.models
 class StageInfo(
     val engines: Int,
     val fuelAmountTons: Double,
-    val burnTimeSec: Double
+    val burnTimeSec: Int
 )
